requirements.txt
streamlit
mysql-connector-python
sshtunnel
pandas
openai
